(base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 % ./verify 请输入待验证的CNF文件名 test.cnf 验证成功！ (base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 % ./main 请输入CNF文件名 test.cnf CNF命题是不可满足的 Time used: 77.6596 ms. (base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 % ./main 请输入CNF文件名 test.cnf ^Z zsh: suspended  ./main (base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 % ./main 请输入CNF文件名 test.cnf CNF命题是可满足的 成真赋值的解在solution文件中。 Time used: 326.429 ms. (base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 % ./verify 请输入待验证的CNF文件名 test.cnf 验证成功！ (base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 % ./main 请输入CNF文件名 test.cnf CNF命题是可满足的 成真赋值的解在solution文件中。 Time used: 210960 ms. (base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 % ./verify 请输入待验证的CNF文件名 test.cnf 验证成功！ (base) xinyiliu@xinyideMacBook-Air 程序设计综合设计课程CNF求解 %

+++ Cnf.hpp
@@ -111,7 +111,7 @@ int Cnf::Delete (int index){
 }
 
 bool Cnf::HaveSingle (void) {
-    long long int i;
+    int i;
     for(i = length - 1; i >= 0; i--){
         if(clauses[i].IsSingle()) return true;
     }
@@ -119,7 +119,7 @@ bool Cnf::HaveSingle (void) {
 }
 
 bool Cnf::HaveEmpty (void) {
-    long long int i;
+    int i;
     for(i = length - 1; i >= 0; i--){
         if(clauses[i].Empty()) return true;
     }
@@ -127,7 +127,7 @@ bool Cnf::HaveEmpty (void) {
 }
 
 int Cnf::FindSingle (void) {
-    long long int i;
+    int i;
     for(i = length - 1; i >= 0; i--){
         if(clauses[i].IsSingle()) return clauses[i].GetFirstLiteral();
     }
@@ -246,3 +246,6 @@ length now is2CNF命题是可满足的
 Time used: 0.681542 ms.
 
 */
+
+
+// 只有在打了一个通宵游戏后才会做出用size_t和longlong去替代int的事情
